requirements.txt
flask
gunicorn
requests
beautifulsoup4
gspread
oauth2client
python-dotenv
gevent
